requirements.txt
PySide6
numpy
matplotlib
networkx
shapely
pyserial
pynmea2
pyproj
schedule
smbus2